+++ src/RetryableJobInterface.php
@@ -1,0 +1,28 @@
+<?php
+/**
+ * @link http://www.yiiframework.com/
+ * @copyright Copyright (c) 2008 Yii Software LLC
+ * @license http://www.yiiframework.com/license/
+ */
+
+namespace yii\queue;
+
+/**
+ * Interface RetryableJobInterface
+ *
+ * @author Roman Zhuravlev <zhuravljov@gmail.com>
+ */
+interface RetryableJobInterface extends JobInterface
+{
+    /**
+     * @return int time to reserve in seconds
+     */
+    public function getTtr();
+
+    /**
+     * @param int $attempt number
+     * @param \Exception|\Throwable $error from last execute of the job
+     * @return bool
+     */
+    public function canRetry($attempt, $error);
+}