+++ src/serializers/SerializerInterface.php
@@ -1,0 +1,30 @@
+<?php
+/**
+ * @link http://www.yiiframework.com/
+ * @copyright Copyright (c) 2008 Yii Software LLC
+ * @license http://www.yiiframework.com/license/
+ */
+
+namespace yii\queue\serializers;
+
+use yii\queue\JobInterface;
+
+/**
+ * Interface SerializerInterface
+ *
+ * @author Roman Zhuravlev <zhuravljov@gmail.com>
+ */
+interface SerializerInterface
+{
+    /**
+     * @param JobInterface|mixed $job
+     * @return string
+     */
+    public function serialize($job);
+
+    /**
+     * @param string $serialized
+     * @return JobInterface
+     */
+    public function unserialize($serialized);
+}