+++ CHANGELOG.md
@@ -4,11 +4,17 @@ Yii2 Queue Extension Change Log
 ## 2.0.1 under development
 
 - Enh #102: Stoppable behavior (zhuravljov)
+- Enh: Rename cli\Verbose behavior to cli\VerboseBehavior (zhuravljov)
+- Enh #147: Igbinary job serializer (xutl)
+- Enh: Rename serializers\Serializer interface to serializers\SerializerInterface (zhuravljov)
+- Enh #122: Rename Job to JobInterface (zhuravljov)
 - Bug #118: Synchronized moving of delayed and reserved jobs to waiting list (zhuravljov)
 - Bug #112: Queue command inside module (tsingsun)
 - Enh #116: Add Chinese Guide (kids-return)
 - Enh #97: Queue::status is public method (zhuravljov)
 - Bug #98: Fixed timeout error handler (zhuravljov)
+- Enh #137: All throwable errors caused by jobs are now caught (brandonkelly)
+- Enh: Added `Signal::setExitFlag()` to stop `Queue::run()` loop manually (silverfire)
 
 ## 2.0.0